feat(MultiOwner): partial Compraventa

+++ Models/Buyer.cs
@@ -9,7 +9,7 @@ namespace RealEstateApp.Models
         [Required]
         public string Rut { get; set; }
         [Required]
-        public int RoyaltyPercentage { get; set; } // % de derecho
+        public double RoyaltyPercentage { get; set; } // % de derecho
         [Required]
         public bool UnaccreditedRoyaltyPercentage { get; set; } // % de derecho no acreditado
         public virtual Inscription Inscription { get; set; }
